micro-api-sdk升级至v1.1.1.r1

+++ src/main/java/com/github/microprograms/yy_vip_center_manager_api/public_api/GoodsBuyLimit_Delete_Api.java
@@ -2,8 +2,11 @@ package com.github.microprograms.yy_vip_center_manager_api.public_api;
 
 import com.github.microprograms.micro_entity_definition_runtime.annotation.Comment;
 import com.github.microprograms.micro_api_runtime.annotation.MicroApiAnnotation;
+import com.github.microprograms.micro_api_runtime.model.Operator;
 import com.github.microprograms.ignite_utils.sql.dml.Condition;
 import com.github.microprograms.ignite_utils.IgniteUtils;
+import com.github.microprograms.micro_api_runtime.exception.MicroApiExecuteException;
+import com.github.microprograms.micro_api_runtime.enums.MicroApiReserveResponseCodeEnum;
 import com.github.microprograms.micro_api_runtime.model.Response;
 import com.github.microprograms.micro_api_runtime.model.Request;
 import com.github.microprograms.micro_api_runtime.utils.MicroApiUtils;
@@ -13,11 +16,20 @@ import com.github.microprograms.micro_entity_definition_runtime.annotation.Requi
 @MicroApiAnnotation(type = "read", version = "v0.0.2")
 public class GoodsBuyLimit_Delete_Api {
 
+    private static Operator<?> getOperator(Req req) {
+        return null;
+    }
+
     private static Object buildFinalCondition(Req req) {
         return Condition.build("id=", req.getBuyLimitId());
     }
 
     private static void core(Req req, Response resp) throws Exception {
+        Operator<?> operator = getOperator(req);
+        if (operator == null)
+            throw new MicroApiExecuteException(MicroApiReserveResponseCodeEnum.unknown_operator_exception);
+        if (operator.isPermissionDenied())
+            throw new MicroApiExecuteException(MicroApiReserveResponseCodeEnum.permission_denied_exception);
         Object finalCondition = buildFinalCondition(req);
         IgniteUtils.deleteObject(GoodsBuyLimit.class, finalCondition);
     }
